move all event stuff under libhoney.events
